+++ app/src/main/java/cityfreqs/com/pilfershushjammer/jammers/ActiveJammerService.java
@@ -95,6 +95,7 @@ public class ActiveJammerService extends Service {
                     CHANNEL_NAME,
                     NotificationManager.IMPORTANCE_LOW);
             channel.setDescription(getResources().getString(R.string.service_state_1));
+            assert notifyManager != null;
             notifyManager.createNotificationChannel(channel);
         }