+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * test Android 10 https://source.android.com/compatibility/android-cdd#5_4_5_concurrent_capture
 * audio source set to VOICE_COMMUNICATION (5.4.5 [C-1-3] silence other app while VOICE_COMM)
 * for API >= 29 setAllowedCapturePolicy to ALLOW_CAPTURE_BY_NONE (for VoIP vs VoIP)
+* added popup toast warning for concurrent audio capture state
 
 ### 4.3.0 (2020-6-13)
 * added headset receiver to passive service for autoswitch inputs